README.md
# CodSoft-CplusplusProgrammer-Internship
I have completed tasks given for my internship as a C++ Programmer in CodSoft. I hope it will be helpfull for you.

Task 1:- Number Guessing Game.
Create a program that generates a random number and asks the
user to guess it. Provide feedback on whether the guess is too
high or too low until the user guesses the correct number.

Task 2:- Simple Calculator
Develop a calculator program that performs basic arithmetic
operations such as addition, subtraction, multiplication, and
division. Allow the user to input two numbers and choose an
operation to perform.

Task 3:- Student Grading System
Create a program that manages student grades. Allow the user
to input student names and their corresponding grades.
Calculate the average grade and display it along with the highest
and lowest grades.