Task 1 Number Game.cpp
# CodSoft-CplusplusProgrammer-Internship
I have completed tasks given for my internship as a C++ Programmer in CodSoft. I hope it will be helpfull for you.

Task 1:- Number Guessing Game.
Create a program that generates a random number and asks the
user to guess it. Provide feedback on whether the guess is too
high or too low until the user guesses the correct number.