Task 3 Student Grading System.cpp
# CodSoft-CplusplusProgrammer-Internship
I have completed tasks given for my internship as a C++ Programmer in CodSoft. I hope it will be helpfull for you.

Task 3:- Student Grading System
Create a program that manages student grades. Allow the user
to input student names and their corresponding grades.
Calculate the average grade and display it along with the highest
and lowest grades.